Refactor the code to seperate UI and logic.
